Fix: Error when inserting an empty list in manyOf
The bug is described in
https://github.com/mdgriffith/elm-markup/issues/42, but I would not
consider it a fix as it does not address the error message helpfullness.

The problem is that in `ExpectManyOf` case of
`Mark.Internal.Description.matchExpected` . There are two variables
there:

    - `oneOptions` `twoOptions`

In case of updates `oneOptions` will be a list of all valid options. The
`twoOptions` will be blocks that are actually inserted.

We want to check if all inserted blocks match any of the valid options.
But it was the other way around - the test was "all valid options are
inserted". Coincidentally when there is only one valid option and at
least one block is inserted, the result is the same. That's probably why
the problem remained undetected for relatively long time. I ran into it
becasue my list was nested in a record (as described in the linked
issue) and was initially empty.